requirements.txt
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.26.0
mysql-connector-python==8.1.1